agent: Attach a bpf_netdev to the main interface automatically
Determine the main interface based on the node address and attach a
bpf_netdev program to it automatically. The behaviour can be disabled by
secifying -d|--device "disabled".

This allows an agent to pick up packets for local endpoints and for
services automatically from both the tunnel mesh and the native network.

Signed-off-by: Thomas Graf <thomas@cilium.io>