agent: Masquerade when tunneling is disabled
We currently only add a masquerade rule for endpoint traffic in overlay
mode. Direct routing mode has been expected to use public IP addresses
where masquerading is not required. The majority of users tend to use
private IPs though which makes endpoints unroutable outside of the
internal network.

A possible solution here is to add a masquerade rule which only applies
if:

 - the source address is within local node prefix
 - the destination address is outside of the cluster and service prefix
 - the packet is routed to the outside of the local node

Fixes: #1164

Signed-off-by: Thomas Graf <thomas@cilium.io>